Move updateAttachmentUploadState out of ChannelLogic
